requirement.txt
streamlit==1.29.0
streamlit-extras==0.3.6
pillow==10.1.0
openai==1.4.0
yfinance==0.2.33
pandas==2.1.4
numpy==1.26.2
langchain==0.0.347
llama-cpp-python

